VMC++ never got adapted to long integers for MCI's number of MC steps. Fix that.

+++ include/vmc/EnergyGradientTargetFunction.hpp
@@ -11,12 +11,12 @@ class EnergyGradientTargetFunction: public nfm::NoisyFunctionWithGradient
 {
 protected:
     VMC &_vmc; // the VMC object containing WaveFunction, Hamiltonian and MCI
-    const int _E_Nmc; // number of MC steps for energy calculation
-    const int _grad_E_Nmc; // number of MC steps for gradient calculation
+    const int64_t _E_Nmc; // number of MC steps for energy calculation
+    const int64_t _grad_E_Nmc; // number of MC steps for gradient calculation
     const double _lambda_reg; // vp regularization factor
 
 public:
-    EnergyGradientTargetFunction(VMC &vmc, int E_Nmc, int grad_E_Nmc, bool useGradErr, double lambda_reg = 0.):
+    EnergyGradientTargetFunction(VMC &vmc, int64_t E_Nmc, int64_t grad_E_Nmc, bool useGradErr, double lambda_reg = 0.):
             nfm::NoisyFunctionWithGradient(vmc.getNVP(), useGradErr), _vmc(vmc),
             _E_Nmc(E_Nmc), _grad_E_Nmc(grad_E_Nmc), _lambda_reg(lambda_reg) {}
