+++ include/vmc/EnergyMinimization.hpp
@@ -14,7 +14,7 @@ namespace vmc
 // Set the Gradient Type to be used via template <> and pass
 // the required constructor arguments.
 template <typename GradT = EnergyGradientTargetFunction>
-void minimizeEnergy(VMC &vmc, nfm::NFM &nfm, int E_NMC, int grad_E_NMC, bool useGradErr = true, double lambda_reg = 0.)
+void minimizeEnergy(VMC &vmc, nfm::NFM &nfm, int64_t E_NMC, int64_t grad_E_NMC, bool useGradErr = true, double lambda_reg = 0.)
 {
     GradT gradfun(vmc, E_NMC, grad_E_NMC, useGradErr, lambda_reg); // create gradient target function of type GradT with passed arguments
     std::vector<double> x0(static_cast<size_t>(vmc.getNVP()));
